rename capnHandler to devnetsHandler
