+++ devnets/mod.ts
@@ -2,10 +2,10 @@
 
 export * from "./binary.ts"
 export * from "./CapiConfig.ts"
-export * from "./capnHandler.ts"
 export * from "./chainSpec.ts"
 export * from "./createTempDir.ts"
 export * from "./DevnetConnection.ts"
+export * from "./devnetsHandler.ts"
 export * from "./proxyWebSocket.ts"
 export * from "./startNetwork.ts"
 export * from "./syncConfig.ts"
